requirements.txt
openai>=1.10.0
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
tqdm>=4.66.0
